feat: add Kali Everywhere section
